requirements.txt
audiomentations==0.42.0
fast_mp3_augment==0.1.0
hydra-core==1.3.2
lameenc==1.8.1
matplotlib==3.10.5
numpy-audio-limiter==0.1.0
pandas==2.3.1
pip==24.0
pretty_midi==0.2.10
pydub==0.25.1
pyloudnorm==0.1.1
pyroomacoustics==0.8.4
pytest==8.4.1
snakeviz==2.2.2
sounddevice==0.5.2
tabulate==0.9.0
torch==2.7.1
tuna==0.5.11
wandb==0.21.0
